+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## 1.0.0
 
-- Initial release with one command to copy `.gitignore` to `search.exclude` and `files.exclude`
+- Initial release with one command to copy `.hgignore` to `search.exclude` and `files.exclude`